requirements-web.txt
# Minimal requirements for web deployment (OpenRouter only)
# Use this for Streamlit Cloud deployment of app_web.py

streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
